--- src/app/shared/models/image.model.ts
@@ -1,11 +1,0 @@
-
-export interface IImageSet {
-  small: IImage;
-  medium?: IImage;
-  large?: IImage;
-}
-
-export interface IImage {
-  src: string;
-  alt?: string;
-}
